# openapi.yaml
openapi: 3.1.0

info:
  title: Web Data Discovery API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Secure geospatial and satellite data discovery platform.
    All endpoints require authentication and are subject to rate limiting.

    ## Rate limiting

    Search endpoints are limited to **300 requests per minute** per client.
    Exceeding the limit returns HTTP 429. Successful responses include
    `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and `X-RateLimit-Reset`.

servers:
  - url: https://api.example.com/v1
    description: Production server (HTTPS only)
  - url: http://localhost:8000/v1
    description: Local development

security:
  - OAuth2:
      - read:data

tags:
  - name: Search
    description: Hybrid satellite and geospatial search

paths:
  /search:
    post:
      operationId: search-data
      tags:
        - Search
      summary: Perform hybrid search across satellite and geospatial data
      description: |
        Executes a hybrid search with metadata/geospatial pre-filtering and
        vector retrieval over curated satellite and geospatial datasets.

        **Rate Limiting**: 300 requests per minute per client.
        Exceeding the limit returns HTTP 429.
      security:
        - OAuth2:
            - read:data
      parameters:
        - name: client_id
          in: query
          required: false
          description: Client identifier used for distributed rate-limit partitioning.
          schema:
            type: string
            default: default
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /search/batch:
    post:
      operationId: batch-search
      tags:
        - Search
      summary: Execute batched hybrid search
      description: |
        Runs multiple search requests in parallel.

        **Rate Limiting**: 300 requests per minute per client (burst 50).
      security:
        - OAuth2:
            - read:data
      parameters:
        - name: client_id
          in: query
          required: false
          description: Client identifier used for distributed rate-limit partitioning.
          schema:
            type: string
            default: default
        - name: per_request_rate_limit
          in: query
          required: false
          description: Apply rate limiting per batch item instead of per batch call.
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 50
              items:
                $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: One result per input request
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BatchSearchResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    OAuth2:
      type: oauth2
      description: OAuth2 Authorization Code flow with PKCE recommended
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/oauth/authorize
          tokenUrl: https://auth.example.com/oauth/token
          scopes:
            read:data: Read access to satellite and geospatial datasets
            write:data: Write access to publish curated data
            admin: Full administrative access

  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current time window
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current time window
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Time when the rate limit resets (Unix timestamp)
      schema:
        type: integer

  schemas:
    SearchRequest:
      type: object
      required:
        - query_vector
      properties:
        query:
          type: string
          maxLength: 500
          default: ""
        query_vector:
          description: Embedding vector (768 or 1024 dimensions).
          oneOf:
            - type: array
              minItems: 768
              maxItems: 768
              items:
                type: number
                format: float
            - type: array
              minItems: 1024
              maxItems: 1024
              items:
                type: number
                format: float
        filters:
          type: object
          additionalProperties: true
          default: {}
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        ef:
          type: integer
          minimum: 1
          maximum: 512
          default: 64
        use_reranking:
          type: boolean
          default: false

    SearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            type: object
            additionalProperties: true
        from_cache:
          type: boolean
          default: false
        degraded:
          type: boolean
          default: false

    BatchSearchResult:
      type: object
      required:
        - request_index
        - results
      properties:
        request_index:
          type: integer
        results:
          type: array
          items:
            type: object
            additionalProperties: true
        from_cache:
          type: boolean
          default: false
        degraded:
          type: boolean
          default: false

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string

    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            additionalProperties: true